chore(si-pkg): break pkg into modules
