+++ lib/si-pkg/.gitignore
@@ -1,1 +1,1 @@
-pkg/
+./pkg/
