+++ lib/si-pkg/src/pkg.rs
@@ -2,7 +2,6 @@ use core::fmt;
 use std::{
     collections::HashMap,
     convert::Infallible,
-    future::Future,
     path::{Path, PathBuf},
     sync::Arc,
 };
@@ -14,14 +13,19 @@ use object_tree::{
 };
 use petgraph::prelude::*;
 use thiserror::Error;
-use url::Url;
+
+mod func;
+mod leaf_function;
+mod prop;
+mod schema;
+mod validation;
+mod variant;
+
+pub use {func::*, leaf_function::*, prop::*, schema::*, validation::*, variant::*};
 
 use crate::{
-    node::{CategoryNode, PkgNode, PropChildNode, PropNode, SchemaVariantChildNode},
-    spec::{
-        FuncArgumentKind, FuncSpecBackendKind, FuncSpecBackendResponseType, LeafInputLocation,
-        LeafKind, PkgSpec, SpecError, ValidationSpecKind,
-    },
+    node::{CategoryNode, PkgNode},
+    spec::{PkgSpec, SpecError},
 };
 
 #[derive(Debug, Error)]
@@ -314,753 +318,5 @@ impl SiPkgMetadata {
 
     pub fn hash(&self) -> Hash {
         self.hash
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SiPkgFuncArgument<'a> {
-    name: String,
-    kind: FuncArgumentKind,
-    element_kind: Option<FuncArgumentKind>,
-
-    hash: Hash,
-    source: Source<'a>,
-}
-
-impl<'a> SiPkgFuncArgument<'a> {
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let hashed_node = &graph[node_idx];
-        let node = match hashed_node.inner() {
-            PkgNode::FuncArgument(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::FUNC_ARGUMENT_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        Ok(Self {
-            name: node.name,
-            kind: node.kind,
-            element_kind: node.element_kind,
-
-            hash: hashed_node.hash(),
-            source: Source::new(graph, node_idx),
-        })
-    }
-
-    pub fn name(&self) -> &str {
-        &self.name
-    }
-
-    pub fn kind(&self) -> FuncArgumentKind {
-        self.kind
-    }
-
-    pub fn element_kind(&self) -> Option<&FuncArgumentKind> {
-        self.element_kind.as_ref()
-    }
-
-    pub fn hash(&self) -> Hash {
-        self.hash
-    }
-
-    pub fn source(&self) -> &Source<'a> {
-        &self.source
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SiPkgFunc<'a> {
-    name: String,
-    display_name: Option<String>,
-    description: Option<String>,
-    handler: String,
-    code_base64: String,
-    backend_kind: FuncSpecBackendKind,
-    response_type: FuncSpecBackendResponseType,
-    hidden: bool,
-    link: Option<Url>,
-    unique_id: Hash,
-
-    hash: Hash,
-    source: Source<'a>,
-}
-
-impl<'a> SiPkgFunc<'a> {
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let func_hashed_node = &graph[node_idx];
-        let func_node = match func_hashed_node.inner() {
-            PkgNode::Func(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::FUNC_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        Ok(Self {
-            name: func_node.name,
-            display_name: func_node.display_name,
-            description: func_node.description,
-            handler: func_node.handler,
-            code_base64: func_node.code_base64,
-            backend_kind: func_node.backend_kind,
-            response_type: func_node.response_type,
-            hidden: func_node.hidden,
-            link: func_node.link,
-            hash: func_hashed_node.hash(),
-            unique_id: func_node.unique_id,
-            source: Source::new(graph, node_idx),
-        })
-    }
-
-    pub fn arguments(&self) -> PkgResult<Vec<SiPkgFuncArgument>> {
-        let mut arguments = vec![];
-        for idx in self
-            .source
-            .graph
-            .neighbors_directed(self.source.node_idx, Outgoing)
-        {
-            arguments.push(SiPkgFuncArgument::from_graph(self.source.graph, idx)?);
-        }
-
-        Ok(arguments)
-    }
-
-    pub fn name(&self) -> &str {
-        self.name.as_ref()
-    }
-
-    pub fn display_name(&self) -> Option<&str> {
-        self.display_name.as_deref()
-    }
-
-    pub fn description(&self) -> Option<&str> {
-        self.description.as_deref()
-    }
-
-    pub fn handler(&self) -> &str {
-        self.handler.as_ref()
-    }
-
-    pub fn code_base64(&self) -> &str {
-        self.code_base64.as_ref()
-    }
-
-    pub fn backend_kind(&self) -> FuncSpecBackendKind {
-        self.backend_kind
-    }
-
-    pub fn response_type(&self) -> FuncSpecBackendResponseType {
-        self.response_type
-    }
-
-    pub fn hidden(&self) -> bool {
-        self.hidden
-    }
-
-    pub fn link(&self) -> Option<&Url> {
-        self.link.as_ref()
-    }
-
-    pub fn hash(&self) -> Hash {
-        self.hash
-    }
-
-    pub fn unique_id(&self) -> Hash {
-        self.unique_id
-    }
-
-    pub fn source(&self) -> &Source<'a> {
-        &self.source
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SiPkgSchema<'a> {
-    name: String,
-    category: String,
-    category_name: Option<String>,
-
-    hash: Hash,
-
-    source: Source<'a>,
-}
-
-impl<'a> SiPkgSchema<'a> {
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let schema_hashed_node = &graph[node_idx];
-        let schema_node = match schema_hashed_node.inner() {
-            PkgNode::Schema(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::SCHEMA_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        let schema = Self {
-            name: schema_node.name,
-            category: schema_node.category,
-            category_name: schema_node.category_name,
-            hash: schema_hashed_node.hash(),
-            source: Source::new(graph, node_idx),
-        };
-
-        Ok(schema)
-    }
-
-    pub fn name(&self) -> &str {
-        self.name.as_ref()
-    }
-
-    pub fn category(&self) -> &str {
-        self.category.as_ref()
-    }
-
-    pub fn category_name(&self) -> Option<&str> {
-        self.category_name.as_deref()
-    }
-
-    pub fn variants(&self) -> PkgResult<Vec<SiPkgSchemaVariant<'a>>> {
-        let mut variants = vec![];
-        for schema_variant_idx in self
-            .source
-            .graph
-            .neighbors_directed(self.source.node_idx, Outgoing)
-        {
-            variants.push(SiPkgSchemaVariant::from_graph(
-                self.source.graph,
-                schema_variant_idx,
-            )?);
-        }
-
-        Ok(variants)
-    }
-
-    pub fn hash(&self) -> Hash {
-        self.hash
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SiPkgSchemaVariant<'a> {
-    name: String,
-    link: Option<Url>,
-    color: Option<String>,
-
-    hash: Hash,
-
-    source: Source<'a>,
-}
-
-impl<'a> SiPkgSchemaVariant<'a> {
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let schema_variant_hashed_node = &graph[node_idx];
-        let schema_variant_node = match schema_variant_hashed_node.inner() {
-            PkgNode::SchemaVariant(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::SCHEMA_VARIANT_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        let schema_variant = Self {
-            name: schema_variant_node.name,
-            link: schema_variant_node.link,
-            color: schema_variant_node.color,
-            hash: schema_variant_hashed_node.hash(),
-            source: Source::new(graph, node_idx),
-        };
-
-        Ok(schema_variant)
-    }
-
-    pub fn name(&self) -> &str {
-        self.name.as_ref()
-    }
-
-    pub fn link(&self) -> Option<&Url> {
-        self.link.as_ref()
-    }
-
-    pub fn color(&self) -> Option<&str> {
-        self.color.as_deref()
-    }
-
-    pub fn leaf_functions(&self) -> PkgResult<Vec<SiPkgLeafFunction>> {
-        let child_idxs = self
-            .source
-            .graph
-            .neighbors_directed(self.source.node_idx, Outgoing)
-            .find(|node_idx| {
-                matches!(
-                    &self.source.graph[*node_idx].inner(),
-                    PkgNode::SchemaVariantChild(SchemaVariantChildNode::LeafFunctions)
-                )
-            })
-            .ok_or(SiPkgError::CategoryNotFound(
-                SchemaVariantChildNode::LeafFunctions.kind_str(),
-            ))?;
-
-        let child_node_idxs: Vec<_> = self
-            .source
-            .graph
-            .neighbors_directed(child_idxs, Outgoing)
-            .collect();
-
-        let mut leaf_functions = vec![];
-        for child_idx in child_node_idxs {
-            leaf_functions.push(SiPkgLeafFunction::from_graph(self.source.graph, child_idx)?);
-        }
-
-        Ok(leaf_functions)
-    }
-
-    fn prop_stack_from_source<I>(
-        source: Source<'a>,
-        node_idx: NodeIndex,
-        parent_id: Option<I>,
-    ) -> PkgResult<Vec<(SiPkgProp, Option<I>)>>
-    where
-        I: Copy,
-    {
-        Ok(
-            match source
-                .graph
-                .neighbors_directed(node_idx, Outgoing)
-                .find(|node_idx| {
-                    matches!(
-                        &source.graph[*node_idx].inner(),
-                        PkgNode::PropChild(PropChildNode::Props)
-                    )
-                }) {
-                Some(prop_child_idxs) => {
-                    let child_node_idxs: Vec<_> = source
-                        .graph
-                        .neighbors_directed(prop_child_idxs, Outgoing)
-                        .collect();
-
-                    let mut entries = vec![];
-                    for child_idx in child_node_idxs {
-                        entries.push((SiPkgProp::from_graph(source.graph, child_idx)?, parent_id));
-                    }
-
-                    entries
-                }
-                None => vec![],
-            },
-        )
-    }
-
-    pub async fn visit_prop_tree<F, Fut, I, C, E>(
-        &'a self,
-        process_prop_fn: F,
-        parent_id: Option<I>,
-        context: &'a C,
-    ) -> Result<(), E>
-    where
-        F: Fn(SiPkgProp<'a>, Option<I>, &'a C) -> Fut,
-        Fut: Future<Output = Result<Option<I>, E>>,
-        E: std::convert::From<SiPkgError>,
-        I: Copy,
-    {
-        let domain_idxs = self
-            .source
-            .graph
-            .neighbors_directed(self.source.node_idx, Outgoing)
-            .find(|node_idx| {
-                matches!(
-                    &self.source.graph[*node_idx].inner(),
-                    PkgNode::SchemaVariantChild(SchemaVariantChildNode::Domain)
-                )
-            })
-            .ok_or(SiPkgError::CategoryNotFound(
-                SchemaVariantChildNode::Domain.kind_str(),
-            ))?;
-
-        let mut child_node_idxs: Vec<_> = self
-            .source
-            .graph
-            .neighbors_directed(domain_idxs, Outgoing)
-            .collect();
-        let domain_node_idx = match child_node_idxs.pop() {
-            Some(idx) => idx,
-            None => Err(SiPkgError::DomainPropNotFound(self.hash()))?,
-        };
-        if !child_node_idxs.is_empty() {
-            Err(SiPkgError::DomainPropMultipleFound(self.hash()))?;
-        }
-
-        // Skip processing the domain prop as a `dal::SchemaVariant` already guarantees such a prop
-        // has already been created. Rather, we will push all immediate children of the domain prop
-        // to be ready for processing.
-        let mut stack =
-            Self::prop_stack_from_source(self.source.clone(), domain_node_idx, parent_id)?;
-
-        while let Some((prop, parent_id)) = stack.pop() {
-            let node_idx = prop.source().node_idx;
-            let new_id = process_prop_fn(prop, parent_id, context).await?;
-
-            stack.extend(Self::prop_stack_from_source(
-                self.source.clone(),
-                node_idx,
-                new_id,
-            )?);
-        }
-
-        Ok(())
-    }
-
-    pub fn hash(&self) -> Hash {
-        self.hash
-    }
-}
-
-#[derive(Clone, Debug)]
-pub struct SiPkgLeafFunction<'a> {
-    func_unique_id: Hash,
-    leaf_kind: LeafKind,
-    inputs: Vec<LeafInputLocation>,
-    hash: Hash,
-    source: Source<'a>,
-}
-
-impl<'a> SiPkgLeafFunction<'a> {
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let hashed_node = &graph[node_idx];
-        let node = match hashed_node.inner() {
-            PkgNode::LeafFunction(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::LEAF_FUNCTION_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        let mut inputs = vec![];
-        if node.input_domain {
-            inputs.push(LeafInputLocation::Domain);
-        }
-        if node.input_resource {
-            inputs.push(LeafInputLocation::Resource);
-        }
-        if node.input_code {
-            inputs.push(LeafInputLocation::Code);
-        }
-        if node.input_deleted_at {
-            inputs.push(LeafInputLocation::DeletedAt);
-        }
-
-        Ok(Self {
-            func_unique_id: node.func_unique_id,
-            leaf_kind: node.leaf_kind,
-            inputs,
-            hash: hashed_node.hash(),
-            source: Source::new(graph, node_idx),
-        })
-    }
-
-    pub fn func_unique_id(&self) -> Hash {
-        self.func_unique_id
-    }
-
-    pub fn leaf_kind(&self) -> LeafKind {
-        self.leaf_kind
-    }
-
-    pub fn inputs(&self) -> &[LeafInputLocation] {
-        &self.inputs
-    }
-
-    pub fn hash(&self) -> Hash {
-        self.hash
-    }
-
-    pub fn source(&self) -> &Source<'a> {
-        &self.source
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum SiPkgValidation<'a> {
-    IntegerIsBetweenTwoIntegers {
-        lower_bound: i64,
-        upper_bound: i64,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    StringEquals {
-        expected: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    StringHasPrefix {
-        expected: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    StringInStringArray {
-        expected: Vec<String>,
-        display_expected: bool,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    StringIsValidIpAddr {
-        hash: Hash,
-        source: Source<'a>,
-    },
-    StringIsHexColor {
-        hash: Hash,
-        source: Source<'a>,
-    },
-    StringIsNotEmpty {
-        hash: Hash,
-        source: Source<'a>,
-    },
-    CustomValidation {
-        func_unique_id: Hash,
-        hash: Hash,
-        source: Source<'a>,
-    },
-}
-
-impl<'a> SiPkgValidation<'a> {
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let hashed_node = &graph[node_idx];
-        let node = match hashed_node.inner() {
-            PkgNode::Validation(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::PROP_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        let hash = hashed_node.hash();
-        let source = Source::new(graph, node_idx);
-
-        Ok(match node.kind {
-            ValidationSpecKind::IntegerIsBetweenTwoIntegers => {
-                SiPkgValidation::IntegerIsBetweenTwoIntegers {
-                    upper_bound: node.upper_bound.ok_or(SiPkgError::ValidationMissingField(
-                        "upper_bound".to_string(),
-                    ))?,
-                    lower_bound: node.lower_bound.ok_or(SiPkgError::ValidationMissingField(
-                        "upper_bound".to_string(),
-                    ))?,
-                    hash,
-                    source,
-                }
-            }
-            ValidationSpecKind::StringEquals => SiPkgValidation::StringEquals {
-                expected: node
-                    .expected_string
-                    .ok_or(SiPkgError::ValidationMissingField(
-                        "expected_string".to_string(),
-                    ))?,
-                hash,
-                source,
-            },
-            ValidationSpecKind::StringHasPrefix => SiPkgValidation::StringHasPrefix {
-                expected: node
-                    .expected_string
-                    .ok_or(SiPkgError::ValidationMissingField(
-                        "expected_string".to_string(),
-                    ))?,
-                hash,
-                source,
-            },
-            ValidationSpecKind::StringInStringArray => {
-                SiPkgValidation::StringInStringArray {
-                    expected: node.expected_string_array.ok_or(
-                        SiPkgError::ValidationMissingField("expected_string_array".to_string()),
-                    )?,
-                    display_expected: node.display_expected.ok_or(
-                        SiPkgError::ValidationMissingField("display_expected".to_string()),
-                    )?,
-                    hash,
-                    source,
-                }
-            }
-
-            ValidationSpecKind::StringIsValidIpAddr => {
-                SiPkgValidation::StringIsValidIpAddr { hash, source }
-            }
-            ValidationSpecKind::StringIsHexColor => {
-                SiPkgValidation::StringIsHexColor { hash, source }
-            }
-            ValidationSpecKind::StringIsNotEmpty => {
-                SiPkgValidation::StringIsNotEmpty { hash, source }
-            }
-            ValidationSpecKind::CustomValidation => {
-                SiPkgValidation::CustomValidation {
-                    func_unique_id: node.func_unique_id.ok_or(
-                        SiPkgError::ValidationMissingField("func_unique_id".to_string()),
-                    )?,
-                    hash,
-                    source,
-                }
-            }
-        })
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum SiPkgProp<'a> {
-    String {
-        name: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    Number {
-        name: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    Boolean {
-        name: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    Map {
-        name: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    Array {
-        name: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-    Object {
-        name: String,
-        hash: Hash,
-        source: Source<'a>,
-    },
-}
-
-impl<'a> SiPkgProp<'a> {
-    pub fn validations(&self) -> PkgResult<Vec<SiPkgValidation>> {
-        Ok(match self {
-            SiPkgProp::Map { source, .. }
-            | SiPkgProp::Array { source, .. }
-            | SiPkgProp::String { source, .. }
-            | SiPkgProp::Number { source, .. }
-            | SiPkgProp::Object { source, .. }
-            | SiPkgProp::Boolean { source, .. } => {
-                let validation_child_idxs = source
-                    .graph
-                    .neighbors_directed(source.node_idx, Outgoing)
-                    .find(|node_idx| {
-                        matches!(
-                            &source.graph[*node_idx].inner(),
-                            PkgNode::PropChild(PropChildNode::Validations)
-                        )
-                    })
-                    .ok_or(SiPkgError::CategoryNotFound(
-                        PropChildNode::Validations.kind_str(),
-                    ))?;
-
-                let child_node_idxs: Vec<_> = source
-                    .graph
-                    .neighbors_directed(validation_child_idxs, Outgoing)
-                    .collect();
-
-                let mut validations = vec![];
-                for child_idx in child_node_idxs {
-                    validations.push(SiPkgValidation::from_graph(source.graph, child_idx)?);
-                }
-
-                validations
-            }
-        })
-    }
-
-    fn from_graph(
-        graph: &'a Graph<HashedNode<PkgNode>, ()>,
-        node_idx: NodeIndex,
-    ) -> PkgResult<Self> {
-        let prop_hashed_node = &graph[node_idx];
-        let prop_node = match prop_hashed_node.inner() {
-            PkgNode::Prop(node) => node.clone(),
-            unexpected => {
-                return Err(SiPkgError::UnexpectedPkgNodeType(
-                    PkgNode::PROP_KIND_STR,
-                    unexpected.node_kind_str(),
-                ))
-            }
-        };
-
-        let hash = prop_hashed_node.hash();
-        let source = Source::new(graph, node_idx);
-
-        Ok(match prop_node {
-            PropNode::String { name } => Self::String { name, hash, source },
-            PropNode::Integer { name } => Self::Number { name, hash, source },
-            PropNode::Boolean { name } => Self::Boolean { name, hash, source },
-            PropNode::Map { name } => Self::Map { name, hash, source },
-            PropNode::Array { name } => Self::Array { name, hash, source },
-            PropNode::Object { name } => Self::Object { name, hash, source },
-        })
-    }
-
-    pub fn name(&self) -> &str {
-        match self {
-            Self::String { name, .. }
-            | Self::Number { name, .. }
-            | Self::Boolean { name, .. }
-            | Self::Map { name, .. }
-            | Self::Array { name, .. }
-            | Self::Object { name, .. } => name,
-        }
-    }
-
-    pub fn hash(&self) -> Hash {
-        match self {
-            Self::String { hash, .. }
-            | Self::Number { hash, .. }
-            | Self::Boolean { hash, .. }
-            | Self::Map { hash, .. }
-            | Self::Array { hash, .. }
-            | Self::Object { hash, .. } => *hash,
-        }
-    }
-
-    fn source(&self) -> &Source<'a> {
-        match self {
-            Self::String { source, .. }
-            | Self::Number { source, .. }
-            | Self::Boolean { source, .. }
-            | Self::Map { source, .. }
-            | Self::Array { source, .. }
-            | Self::Object { source, .. } => source,
-        }
     }
 }
